remove vue 2 support
