+++ .gitignore
@@ -1,26 +1,11 @@
+.cache
 .DS_Store
-node_modules
-/dist
-
-
-# local env files
-.env.local
-.env.*.local
-
-# Log files
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
-
-# Editor directories and files
 .idea
-.vscode
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw?
-
-# npm
-package-lock.json
+*.log
+*.tgz
+coverage
+dist
+lib-cov
+logs
+node_modules
+temp
