+++ package.json
@@ -1,14 +1,57 @@
 {
-    "name": "monorepo",
-    "private": true,
-    "workspaces": [
-      "packages/*"
-    ],
-    "scripts": {
-      "lib:watch": "yarn workspace vue-winbox run build:watch",
-      "lib": "yarn workspace vue-winbox",
-      "demo": "yarn workspace demo",
-      "clean:modules": "rm -rf node_modules && rm -rf packages/lib/node_modules && rm -rf packages/demo/node_modules",
-      "copy:readme": "cp README.md packages/lib/README.md"
+  "name": "vue-winbox",
+  "version": "0.0.0",
+  "packageManager": "pnpm@7.16.1",
+  "description": "",
+  "author": "Robert Soriano <sorianorobertc@gmail.com>",
+  "license": "MIT",
+  "homepage": "https://github.com/wobsoriano/vue-winbox#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/wobsoriano/vue-winbox.git"
+  },
+  "bugs": "https://github.com/wobsoriano/vue-winbox/issues",
+  "keywords": [],
+  "sideEffects": false,
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.js",
+      "import": "./dist/index.mjs"
     }
+  },
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build-fast": "tsup src/index.ts --format cjs,esm",
+    "build": "pnpm build-fast --dts-resolve",
+    "lint": "eslint .",
+    "prepublishOnly": "pnpm build",
+    "release": "bumpp && npm publish",
+    "typecheck": "tsc --noEmit"
+  },
+  "peerDependencies": {
+    "vue": ">=3.2.0"
+  },
+  "dependencies": {
+    "nanoid": "^4.0.0",
+    "winbox": "^0.2.6"
+  },
+  "devDependencies": {
+    "@antfu/eslint-config": "^0.27.0",
+    "@types/winbox": "^0.2.1",
+    "bumpp": "^8.2.1",
+    "eslint": "^8.26.0",
+    "tsup": "6.0.1",
+    "tsx": "^3.10.4",
+    "typescript": "^4.7.4",
+    "vue": "^3.2.45"
+  },
+  "eslintConfig": {
+    "extends": "@antfu"
+  }
 }
