--- packages/demo/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "demo",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build",
-    "serve": "vite preview",
-    "deploy": "yarn build && gh-pages -d dist"
-  },
-  "dependencies": {
-    "vue": "^3.0.5",
-    "vue-winbox": "^0.0.3"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^1.2.5",
-    "@vue/compiler-sfc": "^3.0.5",
-    "gh-pages": "^3.2.3",
-    "typescript": "^4.3.2",
-    "vite": "^2.4.3",
-    "vue-tsc": "^0.0.24"
-  }
-}
