--- packages/lib/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "vue-winbox",
-  "description": "A wrapper component for WinBox.js with the ability to mount Vue components.",
-  "keywords": [
-    "vue",
-    "winbox",
-    "window",
-    "manager"
-  ],
-  "license": "MIT",
-  "version": "0.0.3",
-  "module": "dist/vue-winbox.esm.js",
-  "main": "dist/vue-winbox.cjs.js",
-  "unpkg": "dist/vue-winbox.global.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "rollup -c",
-    "build:watch": "rollup -c -w",
-    "switch:2": "vue-demi-switch 2 vue2",
-    "switch:3": "vue-demi-switch 3"
-  },
-  "dependencies": {
-    "nanoid": "^3.1.23",
-    "vue-demi": "^0.11.2",
-    "winbox": "^0.2.0"
-  },
-  "peerDependencies": {
-    "@vue/composition-api": "^1.0.0-beta.1",
-    "vue": "^2.0.0 || >=3.0.0-rc.0"
-  },
-  "peerDependenciesMeta": {
-    "@vue/composition-api": {
-      "optional": true
-    }
-  },
-  "devDependencies": {
-    "@types/winbox": "^0.2.1",
-    "@vue/compiler-sfc": "^3.0.11",
-    "@vue/composition-api": "^1.0.4",
-    "rollup": "^2.52.8",
-    "rollup-plugin-dts": "^3.0.2",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "tslib": "^2.3.0",
-    "typescript": "^4.2.3",
-    "vue": "^3.0.11"
-  },
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wobsoriano/vue-winbox.git"
-  },
-  "homepage": "https://github.com/wobsoriano/vue-winbox#readme"
-}
