--- packages/lib/shims.vue.d.ts
@@ -1,5 +1,0 @@
-declare module '*.vue' {
-    import type { DefineComponent } from 'vue'
-    const component: DefineComponent<{}, {}, any>
-    export default component
-}
