--- packages/lib/src/composables/useWinBox.ts
@@ -1,5 +1,0 @@
-declare const WinBox: WinBox.WinBoxConstructor;
-
-export default function useWinBox() {
-    return (options: WinBox.Params) => new WinBox(options)
-}
