+++ playground/vite.config.ts
@@ -3,6 +3,5 @@ import vue from '@vitejs/plugin-vue'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue()],
-  base: process.env.NODE_ENV === 'production' ? '/vue-winbox/' : './',
+  plugins: [vue()]
 })
