+++ src/composables/useWinBox.ts
@@ -1,0 +1,5 @@
+declare const WinBox: WinBox.WinBoxConstructor
+
+export function useWinBox() {
+  return (options: WinBox.Params) => new WinBox(options)
+}
