+++ src/index.ts
@@ -1,0 +1,10 @@
+import 'winbox'
+
+import { VueWinBox } from './components/VueWinBox'
+import { useWinBox } from './composables/useWinBox'
+
+export {
+  useWinBox,
+  VueWinBox,
+  VueWinBox as default,
+}
